pyproject.toml
[project]
name = "vision-context-mcp"
version = "0.1.0"
description = "MCP server for image and video analysis with ControlNet preprocessors and 3D entity extraction"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "gdrick"}
]
keywords = ["mcp", "vision", "controlnet", "image-analysis", "video-analysis", "3d"]

# Core dependencies - minimal for basic functionality
dependencies = [
    "mcp>=1.0.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
# Full installation with all models
full = [
    "controlnet-aux>=0.0.7",
    "torch>=2.0.0",
    "torchvision>=0.17.0",
    "transformers>=4.35.0",
    "accelerate>=0.25.0",
    "scipy>=1.11.0",
    "scikit-image>=0.21.0",
    "mediapipe>=0.10.0",
]

# Development tools
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vision_context_mcp"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[project.scripts]
vision-context-mcp = "vision_context_mcp.server:main"